backendpython/requirements.txt
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.38.0
graphviz==0.20.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
numpy==1.24.1
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
sklearn==0.0.post1
threadpoolctl==3.1.0
Werkzeug==2.2.2
